Move grouping logic to library file
